requirements.txt
requests
openai
python-dotenv
schedule
snowflake-connector-python